FreeBSD: Move VmEntryIterator into ptrace module to isolate the unsafe bits
