+++ src/freebsd_maps/mod.rs
@@ -10,8 +10,6 @@ use std::convert::From;
 
 pub type Pid = pid_t;
 
-const FILE_NAME_BUFFER_LENGTH: usize = 4096;
-
 #[derive(Debug, Clone)]
 pub struct MapRange {
     range_start: usize,
@@ -38,90 +36,21 @@ impl MapRange {
     }
 }
 
-impl From<ptrace::vm_entry> for MapRange {
-    fn from(vm_entry: ptrace::vm_entry) -> Self {
-        let pathname = string_from_cstr_ptr(vm_entry.pve_path);
-
+impl From<ptrace::VmEntry> for MapRange {
+    fn from(vm_entry: ptrace::VmEntry) -> Self {
         Self {
             range_start: vm_entry.pve_start as usize,
             range_end: vm_entry.pve_end as usize,
             protection: vm_entry.pve_prot as _,
             offset: vm_entry.pve_offset as usize,
             vnode: vm_entry.pve_fileid as usize,
-            pathname: pathname,
-        }
-    }
-}
-
-#[derive(Default)]
-struct VmEntryIterator {
-    current: c_int,
-    pid: Pid,
-}
-
-impl VmEntryIterator {
-    fn new(pid: Pid) -> std::io::Result<Self> {
-        ptrace::attach(pid)?;
-
-        Ok(Self { current: 0, pid })
-    }
-}
-
-impl Drop for VmEntryIterator {
-    fn drop(&mut self) {
-        ptrace::detach(self.pid);
-    }
-}
-
-impl Iterator for VmEntryIterator {
-    type Item = ptrace::vm_entry;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let Self { current, pid } = *self;
-        // If the region was mapped from a file, `pve_path` contains filename.
-        let pve_pathlen = 4096;
-        let pve_path: [c_char; FILE_NAME_BUFFER_LENGTH] =
-            [0; FILE_NAME_BUFFER_LENGTH];
-
-        let entry = Self::Item {
-            pve_entry: current,
-            pve_path: &pve_path as *const _ as *mut _,
-            pve_pathlen: pve_pathlen,
-            ..Default::default()
-        };
-
-        let result = ptrace::read_vm_entry(pid, entry);
-
-        match result {
-            Ok(entry) => {
-                self.current = entry.pve_entry;
-                Some(entry)
-            }
-            _ => None
-        }
-    }
-}
-
-fn string_from_cstr_ptr(pointer: *const c_char) -> Option<String> {
-    if pointer.is_null() {
-        None
-    } else {
-        unsafe {
-            let result = CStr::from_ptr(pointer)
-                .to_string_lossy()
-                .into_owned();
-
-            if result.len() > 0 {
-                Some(result)
-            } else {
-                None
-            }
+            pathname: vm_entry.pve_path,
         }
     }
 }
 
 pub fn get_process_maps(pid: Pid) -> std::io::Result<Vec<MapRange>> {
-    let iter = VmEntryIterator::new(pid)?;
+    let iter = ptrace::VmEntryIterator::new(pid)?;
 
     Ok(iter.map(MapRange::from).collect())
 }
